Create go build wrapper with naming of final exec based on parent dir
The default behavior of "go build" is to name the final executable after
the first .go source file if files are passed to it. When they are not,
"go build" reverts to naming the the final executable after the current
directory name.

Since I structure my source directories like "<project>/<version>", this
results in a executable named after the version, rather than the
project when I simply run "go build". This isn't what I want...

This wrapper script works around this undesired behavior by
intelligently checking whether source files are being provided and if
not, names the executable after the parent directory name, rather than
the current directory name. Flags can be passed to the wrapper and they
will be forwarded onto the internal "go build" command.

Pretty neat! :^)