Print out moved file with downhere.sh
This is useful for quick and dirty scripting. e.g.

    downhere | xargs open

Which will move the last file in your downloads to the current
directory and open it. Functionally equivalent to

    ls -t1 ~/Downloads | head -1 | xargs -I{} mv {} . && \
        ls -t1 | head -1 | xargs open

This saves you typing that whole thing out, in addition to not having to
ls directories twice due to xargs consuming the filename (though perhaps
there is a better solution).

Pretty neat! :D